+++ go.mod
@@ -2,7 +2,4 @@ module github.com/dondakeshimo/todo-cli
 
 go 1.14
 
-require (
-	github.com/urfave/cli v1.22.4
-	github.com/urfave/cli/v2 v2.2.0 // indirect
-)
+require github.com/urfave/cli/v2 v2.2.0
